Link with gold using LINK_INTERFACE_LIBRARIES for libktorrent.
The error message I got, in a large number of tests (almost all), was:
/opt/kde4/lib/libkdeui.so.5: undefined reference to `QTextBlock::isValid() const'
and it was not fixed by linking to QtGui only.

libktorrent seems to require symbols from libkdeui, so if you link
against libktorrent you always want to link against libkdeui, too.
ld can hide this by pulling in transitive dependencies automatically,
but you shouldn't rely on it.
The proper solution is to state explicitly that libktorrent needs
libkdeui. Due to that, some things can be removed from
target_link_libraries in the CMakeLists.txt of the tests.
KDEUI_LIBS is probably not enough (I'd expect KIO_LIBS), but it is at
least enough to make the tests link.