Report TrackersList statuses statistics for torrent highlighting
Some functionality that reports trackers' statuses to client.
It also allows to distinguish time out errors from others.

The idea is to highlight completed torrents that have no chance to
continue seeding (all trackers are disabled or have error state and
none of which are timeouts) with red. Highlight torrents which
seeding may continue automatically soon (all trackers are with
error or warning state but some errors are timeouts) with some
greenish. And to highlight torrents that supposed to be seeding but
still have some warnings (all tracker states contain warning
messages) with some yellowish. In case at least one tracker reports
OK state with no warning - it will be left green as default.

Differential Revision: https://phabricator.kde.org/D25136