Fix occasional calls to close() with invalid fd -1.
On Kubuntu 24.04, valgrind occasionally shows warnings about close()
being called with an invalid file descriptor of -1. The warnings
point to Socket::close(). The function does check for negative fds,
so it isn't clear how -1 gets to the close() syscall. It is possible
there is some codegen issue. However, reading the fd once before
the check and then passing the read value to the close() syscall
gets rid of the warning.

Additionally, modified Socket constructor and Socket::reset() to
avoid invoking syscalls on invalid fds if creating the socket fails.

Finally, modified Pipe destructor to check for valid fds before
closing them. The fds may end up -1 if the constructor failed to
create the pipe.