Optimize data piece lookup in cache.
Replace the multimap for searching data pieces by chunk with a hash map
that associates chunks with ordered vectors of data pieces. The ordered
vector allows for binary lookup of a matching data piece for a given
chunk. Besides the pointer to the data piece object, vector elements
also contain keys that are composed from the data piece offset and size,
which are used to search for the data piece. Storing the key directly
in the vector helps data cache locality and further speeds up the
search.

In terms of complexity, a data piece lookup in Cache::findPiece has
changed from O(log N), where N is the total number of data pieces in all
chunks, to roughly O(1) + O(log M), where M is the number of data pieces
in the specified chunk. Furthermore, since iterating over a vector is
faster than iterating over a tree, this new data structure also speeds
up other operations that require iteration over data pieces.

I collected a performance profile while KTorrent 23.08.5 was seeding ~40
torrents at about 300-350 Mbit/s on Kubuntu 24.04. Prior to this change,
Cache::findPiece consistently was the second CPU consumer in the profile:

    80.90%  QThread          legacy.so                       [.] RC4
     3.51%  ktorrent         libKF5Torrent.so.23.08.5.abi1   [.] bt::Cache::findPiece(bt::Chunk*, unsigned int, unsigned int, bool)
     2.20%  ktorrent         libc.so.6                       [.] __memmove_avx512_unaligned_erms
     0.63%  ktorrent         libQt5Core.so.5.15.13           [.] QMapNodeBase::nextNode() const
     0.32%  QThread          libQt5Core.so.5.15.13           [.] QMutex::unlock()
     0.29%  ktorrent         libc.so.6                       [.] __memcmp_evex_movbe

After the change, Cache::findPiece has moved to the 57th place:

    85.32%  QThread          legacy.so                       [.] RC4
     3.11%  ktorrent         libc.so.6                       [.] __memmove_avx512_unaligned_erms
     0.31%  ktorrent         libKF5Torrent.so.23.08.5.abi1   [.] bt::CacheFile::map(bt::MMappeable*, unsigned long long, unsigned int, bt::CacheFile::Mode)
     0.25%  ktorrent         libc.so.6                       [.] __memcmp_evex_movbe
     0.25%  QThread          libKF5Torrent.so.23.08.5.abi1   [.] net::UploadThread::waitForSocketsReady()
     0.24%  QThread          libQt5Core.so.5.15.13           [.] QMutex::unlock()
     ...
     0.04%  ktorrent         libKF5Torrent.so.23.08.5.abi1   [.] bt::Cache::findPiece(bt::Chunk*, unsigned int, unsigned int, bool)

In terms of raw performance, it is difficult to produce a definitive
number because of the variable nature of the seeding rate. But
subjectively, KTorrent CPU usage has reduced by ~5%.