Reduced code duplication between Socket constructor and reset().
Moved transport protocol selection to Socket::reset() and extracted
dual-stack and SO_NOSIGPIPE setup to the new Socket::configureFd()
method, which is now used in constructors and reset().

Also, updated the Socket constructor from an fd so that it doesn't
use invalid fd in syscalls, if the Socket object is constructed from
an invalid fd. If the socket is valid, query its type to initialize
the transport protocol, which may be later used in other methods of
Socket (e.g. in reset()).