Avoid using errno in complex expressions in Socket methods.
errno is volatile and may be indirectly modified before it is read for
diagnostics in logging expressions. Furthermore, on Windows it is
redefined to a WSAGetLastError() function call, so reading errno
multiple times results in multiple such calls.

In all Socket methods, read errno once and as early as possible,
before any formatting expressions that could clobber it.