Fix PieceDataInfo constructor, which could dereference a moved-from pointer.
PieceDataInfo constructor that initializes the key from the provided piece
data object used to dereference its argument p, which could be in a moved-from
state after initializing the piece_data member. This would result in a null
dereference.

Use the initialized piece_data member to obtain offset and length to compose
the key.