Fix invalid parsing of UDP tracker scrape response
The code was supposed to read a Uint32 offset from 4th byte of a buffer.
A typo meant that instead of reading from the buffer it was offsetting
from a pointer to the buffer, rather than the data held by the buffer.

Not sure if this fixes any bugs/crashes. I just noticed this when
porting Buffer to be stored as a unique_ptr instead of shared pointer.

Amends 441025c2


(cherry picked from commit 73353cd88d2f50292b260b0de00dbdcc4eb827a1)

Co-authored-by: Jack Hill <jackhill3103@gmail.com>