API/ABI break: Use unique_ptr for Array class
This is safer because it automatically deletes the copy constructor and
assignment operator. Previously these methods were default defined,
which would have resulted in a double delete if someone accidentally
used it.

This is an ABI break but it's fine because this class is not used in
KTorrent or KGet.