API/ABI break: Use Array for internal Buffer data
There's no need to store the Buffer data in a shared_ptr. Ownership is
traded between the Buffer and BufferPool only when the Buffer is created
and destroyed.

Also add the move constructor and assignment operator to Array. Both
methods will set the capacity of the moved-from array to 0, to avoid
unneccessary allocations (in this specific case we are not interested in
re-using the old Array object).

This is an API and ABI break but it's fine because these classes are not
used in KTorrent or KGet.