Improve superseeder alternative chunk selection, also fixes the test
The point of superseeding is to minimize the time taken to distribute an
entire (usually new) torrent into a swarm.

Effectively each peer of a superseeder will only be allowed to download
one piece at a time, and the piece is chosen such that no other peers
currently have it. Then the peers will take it upon themselves to share
these pieces to each other.

If all of the pieces have been shared at least once, then the
superseeder will start to share only the rarest pieces, i.e. pieces that
only the smallest number of peers have.

The previous code was correctly preferring to send pieces that were not
owned by any other peers, but it was not taking into account the fact
that we may have already shared that piece and are waiting for a
different peer to confirm that it has it.

Adding the check for active_chunks mitigates this issue and ensures that
we only send unique pieces until we've shared to the point that all
chunks are in circulation.

Closes #3