API/ABI break: Port BDecoder to take a QByteArrayView when parsing
There's not really any need to do a deep copy of this data. Instead one
can use the `toByteArray()` and `toString()` methods of `bt::Value` to
get a copy of the individual values when needed. Or one can access the
raw bencoded data of the value with `BNode::getData()` and create a copy
manually.

Previously the code was a bit awkward. When passing a QByteArray, it
would always take a deep copy subject to COW rules. I.e. if someone
modified the data from outside the class, it would cause a deep copy.
The behaviour was different when passing just a pointer since
fromRawData will treat the data as a view.

Both constructors did a deep copy anyway since we use operator[]
everywhere. Using QByteArrayView means we can avoid this.

Also remove the `offset` parameter. It's pointless when one can just
pass a sliced QByteArrayView.

No downstream changes required as KTorrent and KGet master branches have
already been ported to use the default value of the offset parameter
instead of passing it explicitely.