Fix crash when killing DataCheckerJob
The `kill()` function will cancel the thread, wait for it to finish,
call `deleteLater()`, then set the thread to nullptr. Then on the next
iteration of the main loop, the slot `threadFinished()` is called, where
we try to call `deleteLater()` on the thread again, even though the
thread is now set to nullptr.

Instead only delete the finished thread if it has not been killed.

Fixes sentry issue KTORRENT-WJ